byte-int disaseter
Due to my error with byte being -126-127 instead of 0-255, moved all internal work to int, and retyped it to byte only after jsut before sending
Also delayed BT sending byte by byte and with delay of 1. but it looks like not beeing the sync problem
most suspected is 2A instead 1Amper

+++ src/org/fbb/board/internals/Boulder.java
@@ -93,7 +93,7 @@ public class Boulder implements Cloneable {
         }
     }
 
-    void apply(byte[] psStatus, int heigt) {
+    void apply(int[] psStatus, int heigt) {
         String start = map.getProperty(START);
         String path = map.getProperty(PATH);
         String top = map.getProperty(TOP);
@@ -109,7 +109,7 @@ public class Boulder implements Cloneable {
                 && map.getProperty(PATH).trim().isEmpty();
     }
 
-    private void apply(byte[] values, String compressedValue, byte mark, int heigt) {
+    private void apply(int[] values, String compressedValue, int mark, int heigt) {
         if (compressedValue == null || compressedValue.trim().isEmpty()) {
             return;
         }
